chatbot-multicanal/requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langchain-experimental==0.0.47
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0
aiohttp==3.9.1
twilio==8.10.0
python-telegram-bot==20.7
flask==3.0.0
redis==5.0.1
chromadb==0.4.18

